+++ internal/js_parser/js_parser.go
@@ -6357,16 +6357,15 @@ const (
 func (p *parser) parseDecorators(decoratorScope *js_ast.Scope, classKeyword logger.Range, context decoratorContextFlags) (decorators []js_ast.Decorator) {
 	if p.lexer.Token == js_lexer.TAt {
 		if p.options.ts.Parse {
-			if (context & decoratorInClassExpr) != 0 {
+			if (context&decoratorInFnArgs) != 0 && p.options.ts.Config.ExperimentalDecorators != config.True {
+				p.log.AddErrorWithNotes(&p.tracker, p.lexer.Range(), "Parameter decorators only work when experimental decorators are enabled", []logger.MsgData{{
+					Text: "You can enable experimental decorators by adding \"experimentalDecorators\": true to your \"tsconfig.json\" file.",
+				}})
+			} else if (context & decoratorInClassExpr) != 0 {
 				p.lexer.AddRangeErrorWithNotes(p.lexer.Range(), "Experimental decorators can only be used with class declarations in TypeScript",
 					[]logger.MsgData{p.tracker.MsgData(classKeyword, "This is a class expression, not a class declaration:")})
 			} else if (context & decoratorBeforeClassExpr) != 0 {
 				p.log.AddError(&p.tracker, p.lexer.Range(), "Experimental decorators cannot be used in expression position in TypeScript")
-			} else if p.options.ts.Config.ExperimentalDecorators != config.True {
-				p.log.AddErrorWithNotes(&p.tracker, p.lexer.Range(), "Experimental decorators are not currently enabled", []logger.MsgData{{
-					Text: "To use experimental decorators in TypeScript with esbuild, you need to enable them by adding \"experimentalDecorators\": true in your \"tsconfig.json\" file. " +
-						"TypeScript's experimental decorators are currently the only kind of decorators that esbuild supports.",
-				}})
 			}
 		} else {
 			if (context & decoratorInFnArgs) != 0 {