+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-* Pass through JavaScript decorators in JavaScript files ([#104](https://github.com/evanw/esbuild/issues/104))
+* Pass through JavaScript decorators ([#104](https://github.com/evanw/esbuild/issues/104))
 
     In this release, esbuild now parses decorators from the upcoming [JavaScript decorators proposal](https://github.com/tc39/proposal-decorators) and passes them through to the output unmodified, at least as long as the language target is set to `esnext`. Transforming JavaScript decorators to environments that don't support them has not been implemented yet.